backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
Pillow>=10.0
python-decouple>=3.8
django-cors-headers>=4.3